[not verified] Split forms inbox view into multiple components/files
